# ui/openapi.yaml
openapi: 3.1.0
info:
  title: Web Gagebu API
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /api/v1/account-books/{accountBookId}/members/{memberId}:
    get:
      tags:
        - Member
      summary: Get member by ID
      operationId: getMember
      parameters:
        - name: accountBookId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MemberView'
        default:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - cookieTokenAuth: []
        - bearerTokenAuth: []
    put:
      tags:
        - Member
      summary: Update member role
      operationId: updateMember
      parameters:
        - name: accountBookId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdateRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MemberView'
        default:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - cookieTokenAuth: []
        - bearerTokenAuth: []
  /api/v1/account-books/{accountBookId}/categories:
    get:
      tags:
        - Category
      summary: Get categories for account book
      operationId: getCategories
      parameters:
        - name: accountBookId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        default:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - cookieTokenAuth: []
        - bearerTokenAuth: []
    put:
      tags:
        - Category
      summary: Replace categories for account book
      description: "If a category is not included in the request, it will be deleted.\
        \ New categories can be added by providing a null ID."
      operationId: updateCategories
      parameters:
        - name: accountBookId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryListUpdateRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        default:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - cookieTokenAuth: []
        - bearerTokenAuth: []
...
  /api/v1/account-books/{accountBookId}/members:
    get:
      tags:
        - Member
      summary: List members
      operationId: listMembers
      parameters:
        - name: accountBookId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MemberListResponse'
        default:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - cookieTokenAuth: []
        - bearerTokenAuth: []
...
components:
  schemas:
...
    CategoryView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Hobby
        isBasic:
          type: boolean
        isFallback:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MemberView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - OWNER
            - PARTICIPANT
        nickname:
          type: string
          example: Marilly
        pictureUrl:
          type: string
          example: https://example.com/pic.png
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MemberUpdateRequest:
      type: object
      properties:
        role:
          type: string
          enum:
            - OWNER
            - PARTICIPANT
    CategoryListUpdateRequest:
      type: object
      properties:
        categories:
          type: array
          description: List of categories to update
          items:
            $ref: '#/components/schemas/CategoryUpdateRequest'
      required:
        - categories
    CategoryUpdateRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Category ID - null for new categories
        name:
          type: string
          description: Category name
          example: Food & Dining
    CategoryListResponse:
      type: object
      properties:
        categories:
          type: array
          description: List of categories
          items:
            $ref: '#/components/schemas/CategoryView'
...
    MemberListResponse:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberView'
...
    ErrorResponse:
      description: Common error response structure
      properties:
        status:
          type: integer
          format: int32
          example: 404
        statusText:
          type: string
          example: Not Found
        errorCode:
          type: string
          description: Detailed error codes for the application
          enum:
            - BAD_REQUEST
            - UNEXPECTED_IDP
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - UNPROCESSABLE_ENTITY
            - INTERNAL_SERVER_ERROR
            - NOT_IMPLEMENTED
            - SERVICE_UNAVAILABLE
          example: DETAILED_ERROR_CODE
        message:
          type: string
          example: Something went wrong
        requestId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
  securitySchemes:
    bearerTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieTokenAuth:
      type: apiKey
      name: TOKEN
      in: cookie